requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
rich
requests
python-multipart
pydantic
pdfplumber